+++ packages/metal-icons/src/16/outline/CircleArrowRight.js
@@ -1,0 +1,17 @@
+import * as React from 'react'
+const SvgCircleArrowRight = (props) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={16}
+    height={16}
+    fill="currentColor"
+    {...props}
+  >
+    <path d="M8.245 4.695a.75.75 0 0 0-.05 1.06l1.36 1.495H4.75a.75.75 0 0 0 0 1.5h4.805l-1.36 1.495a.75.75 0 0 0 1.11 1.01l2.5-2.75a.75.75 0 0 0 0-1.01l-2.5-2.75a.75.75 0 0 0-1.06-.05Z" />
+    <path
+      fillRule="evenodd"
+      d="M0 8a8 8 0 1 1 16 0A8 8 0 0 1 0 8m8-6.5a6.5 6.5 0 1 0 0 13 6.5 6.5 0 0 0 0-13"
+    />
+  </svg>
+)
+export default SvgCircleArrowRight