+++ packages/metal-icons/src/16/outline/CircleCaretDown.js
@@ -1,0 +1,17 @@
+import * as React from 'react'
+const SvgCircleCaretDown = (props) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={16}
+    height={16}
+    fill="currentColor"
+    {...props}
+  >
+    <path d="M10.908 6.39a.75.75 0 0 1-.027.766l-2.25 3.5a.75.75 0 0 1-1.262 0l-2.25-3.5A.75.75 0 0 1 5.749 6h4.5a.75.75 0 0 1 .659.39" />
+    <path
+      fillRule="evenodd"
+      d="M8 0a8 8 0 1 1 0 16A8 8 0 0 1 8 0m0 1.5a6.5 6.5 0 1 1 0 13 6.5 6.5 0 0 1 0-13"
+    />
+  </svg>
+)
+export default SvgCircleCaretDown