+++ packages/metal-icons/src/16/solid/CircleCaretLeft.js
@@ -1,0 +1,16 @@
+import * as React from 'react'
+const SvgCircleCaretLeft = (props) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={16}
+    height={16}
+    fill="currentColor"
+    {...props}
+  >
+    <path
+      fillRule="evenodd"
+      d="M16 8A8 8 0 1 1-.001 8 8 8 0 0 1 16 8m-6.39 2.908a.75.75 0 0 1-.766-.027l-3.5-2.25a.75.75 0 0 1 0-1.262l3.5-2.25A.75.75 0 0 1 10 5.75v4.5a.75.75 0 0 1-.39.658"
+    />
+  </svg>
+)
+export default SvgCircleCaretLeft