+++ packages/metal-icons/src/16/solid/CircleCaretUp.js
@@ -1,0 +1,16 @@
+import * as React from 'react'
+const SvgCircleCaretUp = (props) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={16}
+    height={16}
+    fill="currentColor"
+    {...props}
+  >
+    <path
+      fillRule="evenodd"
+      d="M8 16A8 8 0 1 1 8 0a8 8 0 0 1 0 16M5.092 9.61a.75.75 0 0 1 .027-.766l2.25-3.5a.75.75 0 0 1 1.262 0l2.25 3.5A.75.75 0 0 1 10.25 10h-4.5a.75.75 0 0 1-.658-.39"
+    />
+  </svg>
+)
+export default SvgCircleCaretUp