+++ packages/metal-icons/src/24/solid/CircleCaretDown.js
@@ -1,0 +1,16 @@
+import * as React from 'react'
+const SvgCircleCaretDown = (props) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={24}
+    height={24}
+    fill="currentColor"
+    {...props}
+  >
+    <path
+      fillRule="evenodd"
+      d="M1 12C1 5.925 5.925 1 12 1s11 4.925 11 11-4.925 11-11 11S1 18.075 1 12m10.426 3.709c.258.388.888.388 1.146 0l3.333-4.808c.265-.397-.056-.901-.573-.901H8.668c-.517 0-.838.504-.573.901l3.33 4.808Z"
+    />
+  </svg>
+)
+export default SvgCircleCaretDown